+++ pytest/test_hydrostatics.py
@@ -5,7 +5,6 @@ from pathlib import Path
 
 import pytest
 
-import pandas as pd
 import numpy as np
 import capytaine as cpt