+++ src/capytaine/bem/engines.py
@@ -9,7 +9,8 @@ from typing import Tuple, Union, Optional, Callable
 import numpy as np
 import scipy.sparse.linalg as ssl
 
-from capytaine.meshes.symmetric import ReflectionSymmetricMesh
+from capytaine.meshes.symmetric import ReflectionSymmetricMesh as OldReflectionSymmetricMesh
+from capytaine.new_meshes.symmetric_meshes import ReflectionSymmetricMesh, RotationSymmetricMesh
 
 from capytaine.green_functions.abstract_green_function import AbstractGreenFunction
 from capytaine.green_functions.delhommeau import Delhommeau
@@ -124,77 +125,87 @@ class BasicMatrixEngine(MatrixEngine):
     def _repr_pretty_(self, p, cycle):
         p.text(self.__str__())
 
-    def build_S_matrix(
-            self, mesh1, mesh2, free_surface, water_depth, wavenumber
-            ) -> np.ndarray:
+    def build_S_matrix(self, mesh1, mesh2, **gf_params) -> np.ndarray:
         """Similar to :code:`build_matrices`, but returning only :math:`S`"""
         # Calls directly evaluate instead of build_matrices because the caching
         # mechanism of build_matrices is not compatible with giving mesh1 as a
         # list of points, but we need that for post-processing
-        S, _ = self.green_function.evaluate(
-                mesh1, mesh2, free_surface, water_depth, wavenumber
-            )
+        S, _ = self.green_function.evaluate(mesh1, mesh2, **gf_params)
         return S
 
-    def build_fullK_matrix(
-            self, mesh1, mesh2, free_surface, water_depth, wavenumber
-            ) -> np.ndarray:
+    def build_fullK_matrix(self, mesh1, mesh2, **gf_params) -> np.ndarray:
         """Similar to :code:`build_matrices`, but returning only full :math:`K`
         (that is the three components of the gradient, not just the normal one)"""
         # TODO: could use symmetries. In particular for forward, we compute the
         # full velocity on the same mesh so symmetries could be used.
-        _, fullK = self.green_function.evaluate(
-                mesh1, mesh2, free_surface, water_depth, wavenumber,
-                adjoint_double_layer=True, early_dot_product=False,
-            )
+        gf_params.setdefault("diagonal_term_in_double_layer", True)
+        gf_params.setdefault("adjoint_double_layer", True)
+        gf_params.setdefault("early_dot_product", False)
+        _, fullK = self.green_function.evaluate(mesh1, mesh2, **gf_params)
         return fullK
 
-    def _build_matrices_with_symmetries(
-            self, mesh1, mesh2, *, free_surface, water_depth, wavenumber, adjoint_double_layer
-            ) -> Tuple[MatrixLike, MatrixLike]:
-        if (isinstance(mesh1, ReflectionSymmetricMesh)
-                and isinstance(mesh2, ReflectionSymmetricMesh)
+    def _build_matrices_with_symmetries(self, mesh1, mesh2, *, diagonal_term_in_double_layer=True, **gf_params) -> Tuple[MatrixLike, MatrixLike]:
+        if (isinstance(mesh1, OldReflectionSymmetricMesh)
+                and isinstance(mesh2, OldReflectionSymmetricMesh)
                 and mesh1.plane == mesh2.plane):
 
-            S_a, K_a = self._build_matrices_with_symmetries(
-                mesh1[0], mesh2[0],
-                free_surface=free_surface, water_depth=water_depth,
-                wavenumber=wavenumber, adjoint_double_layer=adjoint_double_layer
-                )
-            S_b, K_b = self._build_matrices_with_symmetries(
-                mesh1[0], mesh2[1],
-                free_surface=free_surface, water_depth=water_depth,
-                wavenumber=wavenumber, adjoint_double_layer=adjoint_double_layer
-                )
+            S_a, K_a = self._build_matrices_with_symmetries(mesh1[0], mesh2[0],
+                                                            diagonal_term_in_double_layer=diagonal_term_in_double_layer, **gf_params)
+            S_b, K_b = self._build_matrices_with_symmetries(mesh1[0], mesh2[1],
+                                                            diagonal_term_in_double_layer=False, **gf_params)
 
             return BlockCirculantMatrix([S_a, S_b]), BlockCirculantMatrix([K_a, K_b])
 
+        elif (isinstance(mesh1, ReflectionSymmetricMesh)
+                and isinstance(mesh2, ReflectionSymmetricMesh)
+                and mesh1.plane == mesh2.plane):
+
+            S_a, K_a = self._build_matrices_with_symmetries(mesh1.half, mesh2.half,
+                                                            diagonal_term_in_double_layer=diagonal_term_in_double_layer, **gf_params)
+            S_b, K_b = self._build_matrices_with_symmetries(mesh1.other_half, mesh2.half,
+                                                            diagonal_term_in_double_layer=False, **gf_params)
+
+            return BlockCirculantMatrix([S_a, S_b]), BlockCirculantMatrix([K_a, K_b])
+
+        elif (isinstance(mesh1, RotationSymmetricMesh)
+                and isinstance(mesh2, RotationSymmetricMesh)
+                and mesh1.n == mesh2.n):
+
+            S_cols, K_cols = self.green_function.evaluate(
+                    mesh1.merged(), mesh2.wedge,
+                    diagonal_term_in_double_layer=diagonal_term_in_double_layer,
+                    **gf_params,
+                    )
+            # Building the first column of blocks, that is the interactions of all of mesh1 with the reference wedge of mesh2.
+
+            n_blocks = mesh1.n # == mesh2.n
+            block_shape = (mesh2.wedge.nb_faces, mesh2.wedge.nb_faces)
+
+            return (
+                    BlockCirculantMatrix(S_cols.reshape((n_blocks, *block_shape))),
+                    BlockCirculantMatrix(K_cols.reshape((n_blocks, *block_shape))),
+                    )
+
         else:
-            return self.green_function.evaluate(
-                mesh1, mesh2, free_surface, water_depth, wavenumber,
-                adjoint_double_layer=adjoint_double_layer, early_dot_product=True,
-            )
+            gf_params.setdefault("early_dot_product", True)
+            return self.green_function.evaluate(mesh1, mesh2, diagonal_term_in_double_layer=diagonal_term_in_double_layer, **gf_params)
 
     def _build_and_cache_matrices_with_symmetries(
-            self, mesh1, mesh2, *, free_surface, water_depth, wavenumber, adjoint_double_layer
+            self, mesh1, mesh2, **gf_params
             ) -> Tuple[MatrixLike, LUDecomposedMatrixOrNot]:
-        if (mesh1, mesh2, free_surface, water_depth, wavenumber, adjoint_double_layer) == self.last_computed_inputs:
+        if (mesh1, mesh2, gf_params) == self.last_computed_inputs:
             LOG.debug("%s: reading cache.", self.__class__.__name__)
             return self.last_computed_matrices
         else:
             LOG.debug("%s: computing new matrices.", self.__class__.__name__)
             self.last_computed_matrices = None  # Unlink former cached values, so the memory can be freed to compute new matrices.
-            S, K = self._build_matrices_with_symmetries(
-                    mesh1, mesh2,
-                    free_surface=free_surface, water_depth=water_depth,
-                    wavenumber=wavenumber, adjoint_double_layer=adjoint_double_layer
-                    )
-            self.last_computed_inputs = (mesh1, mesh2, free_surface, water_depth, wavenumber, adjoint_double_layer)
+            S, K = self._build_matrices_with_symmetries(mesh1, mesh2, **gf_params)
+            self.last_computed_inputs = (mesh1, mesh2, gf_params)
             self.last_computed_matrices = (S, K)
             return self.last_computed_matrices
 
     # Main interface for compliance with AbstractGreenFunction interface
-    def build_matrices(self, mesh1, mesh2, free_surface, water_depth, wavenumber, adjoint_double_layer=True):
+    def build_matrices(self, mesh1, mesh2, **gf_params):
         r"""Build the influence matrices between mesh1 and mesh2.
 
         Parameters
@@ -218,9 +229,7 @@ class BasicMatrixEngine(MatrixEngine):
             the matrices :math:`S` and :math:`K`
         """
         return self._build_and_cache_matrices_with_symmetries(
-            mesh1, mesh2,
-            free_surface=free_surface, water_depth=water_depth,
-            wavenumber=wavenumber, adjoint_double_layer=adjoint_double_layer
+            mesh1, mesh2, **gf_params
         )
 
     def linear_solver(self, A: LUDecomposedMatrixOrNot, b: np.ndarray) -> np.ndarray:
@@ -250,7 +259,7 @@ class BasicMatrixEngine(MatrixEngine):
             return x
 
         elif self._linear_solver in ("lu_decomposition", "lu_decomposition_with_overwrite") :
-            overwrite_a = (self._linear_solver == "lu_decompositon_with_overwrite")
+            overwrite_a = (self._linear_solver == "lu_decomposition_with_overwrite")
             if not has_been_lu_decomposed(A):
                 luA = lu_decompose(A, overwrite_a=overwrite_a)
                 if A is self.last_computed_matrices[1]:
@@ -270,11 +279,11 @@ class BasicMatrixEngine(MatrixEngine):
             raise NotImplementedError(
                 f"Unknown `linear_solver` in BasicMatrixEngine: {self._linear_solver}"
             )
-        
+
     def compute_ram_estimation(self, problem):
         nb_faces = problem.body.mesh.nb_faces
         nb_matrices = 2
-        nb_bytes = 16 
+        nb_bytes = 16
 
         if self._linear_solver == "lu_decomposition":
             nb_matrices += 1
@@ -282,5 +291,43 @@ class BasicMatrixEngine(MatrixEngine):
         if self.green_function.floating_point_precision == "float32":
             nb_bytes = 8
 
-        peak_memory = nb_faces**2*nb_matrices*nb_bytes/1e9
-        return peak_memory
+        # In theory a simple symmetry is a gain of factor 1/2
+        # and a nested symmetry is a gain of factor 1/4.
+        # For the solvers that use LU decomposition the gain is a bit less.
+        solver_factors = {
+            # Formula to compute the factor of gain:
+            # (2 matrices * theoretical symmetry factor + LU decomposition + intermediate_step) / nb matrices without symmetry
+            "lu_decomposition": {
+                "simple": 2 / 3, # (2 * 1/2 + 1/2 + 1/2) / 3
+                "nested": 5 / 12,  # (2 * 1/4 + 1/4 + 1/2) / 3
+                "rotation": 4 / 3,
+            },
+            # Formula to compute the factor of gain:
+            # (2 matrices * theoretical symmetry factor + intermediate step) / nb matrices without symmetry
+            "lu_decomposition_with_overwrite": {
+                "simple": 3 / 4, # (2 * 1/2 + 1/2) / 2
+                "nested": 1 / 2, # (2 * 1/4 + 1/2) / 2
+                "rotation": 3 / 2,
+            },
+            "gmres": {
+                "simple": 1 / 2,
+                "nested": 1 / 4,
+                "rotation": 1,
+            },
+        }
+
+        if isinstance(problem.body.mesh, ReflectionSymmetricMesh):
+            if isinstance(problem.body.mesh.half, ReflectionSymmetricMesh):
+                # Should not go deeper than that, there is currently only two
+                # symmetries available
+                symmetry_type = "nested"
+            else:
+                symmetry_type = "simple"
+            symmetry_factor = solver_factors[self._linear_solver][symmetry_type]
+        elif isinstance(problem.body.mesh, RotationSymmetricMesh):
+            symmetry_factor = solver_factors[self._linear_solver]["rotation"] / problem.body.mesh.n
+        else:
+            symmetry_factor = 1.0
+
+        memory_peak = symmetry_factor * nb_faces**2 * nb_matrices * nb_bytes/1e9
+        return memory_peak