+++ docs/changelog.rst
@@ -70,6 +70,12 @@ Internals
 
 * Move hydrostatics routines in a dedicated module and rewrite corresponding tests (:pull:`794`)
 
+* Refactor the implementation of the timer to make it easier to include more steps (:pull:`809`)
+
+* Parameters ``free_surface``, ``water_depth`` and ``wavenumber`` are always
+  keyword arguments in the engine and the Green function (:pull:`812`)
+
+
 ---------------------------------
 New in version 2.3.1 (2025-10-14)
 ---------------------------------